:art: move SideCategoryList
