+++ src/components/navigations/stories/SideCategoryList.stories.tsx
@@ -1,4 +1,4 @@
-import SideCategoryList from '@components/navigations/SideCategoryList';
+import SideCategoryList from '@components/navigations/side-category-list';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
